+++ apps/api/src/app.dto.ts
@@ -22,6 +22,6 @@ export class AppMetaResponseDto {
   @ApiProperty({ example: '2026-01-09T15:54:13.000Z', nullable: true })
   buildTime!: string | null;
 
-  @ApiProperty({ example: '/recommendations' })
+  @ApiProperty({ example: '/immaculaterr' })
   publicBasePath!: string;
 }